requirements.txt
#
# Add packages here to install them using:
#   pip install -r requirements.txt
#

-r requirements/conda.txt

Flask-Script
vcf-parser
